[refactor] 관심사 음성 기반 동화 생성 로직 FastApiStoryClient 내부에 삽입
